lab 12 added authentication

+++ routes/tasks.js
@@ -2,10 +2,19 @@ var express = require('express');
 var router = express.Router();
 var Task = require('../models/task');
 
+function isLoggedIn(req, res, next) {
+    if (req.isAuthenticated()) {
+        res.locals.username = req.user.local.username;
+        next();
+    } else {
+        res.redirect('/auth')
+    }
+}
+router.use(isLoggedIn);
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
-  Task.find({completed: false})
+  Task.find({_creator: req.user, completed: false})
       .then((docs) => {
           res.render('index', { title: 'Incomplete tasks', tasks: docs});
       })
@@ -15,44 +24,47 @@ router.get('/', function(req, res, next) {
 });
 /*Add a new task*/
 router.post('/add', function(req, res, next){
-    var newDate = new Date();
-   var t = new Task({text:req.body.text, completed: false, dateCreated: newDate});
-    if(req.body.text){
-        t.save().then((newTask) => {
-            console.log('The new task created is ', newTask);
-            res.redirect('/');
-        }).catch(() => {
-            next(err);
-        });
-    }
-   else{
-        req.flash('error', 'Please enter a task.');
+
+    if (!req.body || !req.body.text) {
+        req.flash('error', 'Please enter some text');
         res.redirect('/');
+    }
+
+    else {
+        // Save new task with text provided, for the current user, and completed = false
+        var task = Task({ _creator: req.user, text : req.body.text, completed: false});
+
+        task.save()
+            .then(() => {
+                res.redirect('/');
+            })
+            .catch((err) => {
+                next(err);
+            });
     }
 });
 /*finish the task and send to completed tasks*/
 router.post('/done', function(req, res, next){
-    var newDate = new Date();
-  Task.findByIdAndUpdate(req.body._id, {completed:true, dateCompleted: newDate})
-      .then((originalTask) => {
-    if(originalTask){
-        res.redirect('/');
-        req.flash('info', originalTask.text, ' marked as done!');
-    }else{
-        var err = new Error('Not Found');
-        err.status = 404;
-        console.log(err);
-        next(err);
-    }
-  })
-      .catch((err) => {
-    next(err);
-    console.log(err);
-      })
+    Task.findOneAndUpdate( {_id: req.body._id, _creator: req.user.id}, {completed: true})
+        .then( (task) => {
+
+            if (!task) {
+                res.status(403).send('This is not your task!');
+            }
+
+            else {
+                req.flash('info', 'Task marked as done');
+                res.redirect('/')
+            }
+
+        })
+        .catch( (err) => {
+            next(err);
+        });
 });
 /*completed page*/
 router.get('/completed', function(req, res, next) {
-    Task.find({completed:true})
+    Task.find({_creator: req.user._id, completed:true})
         .then((docs) => {
             res.render('completed_tasks', {tasks: docs});
         }).catch((err) => {
@@ -61,64 +73,66 @@ router.get('/completed', function(req, res, next) {
 });
 /*delete the specific tasks*/
 router.post('/delete', function(req, res, next){
-    console.log(req.body);
-  Task.findByIdAndRemove(req.body._id)
-      .then((deletedTask) => {
-    if(deletedTask){
-        res.redirect('/');
-        req.flash('info', 'Task deleted');
-    } else{
-      var error = new Error('Task not found');
-      error.status = 404;
-      console.log(err);
-      next(error);
-    }
-      })
-      .catch((err) => {
-          var error = new Error('Task not found');
-          error.status = 404;
-          console.log(err);
-          next(err);
-      });
+
+    Task.findOneAndRemove( {_id: req.body._id, _creator: req.user.id}, {completed: true} )
+        .then( (task) => {
+            if (!task)  { // No task deleted, therefore the ID is not valid,
+                //or did not belong to the current logged in user.
+                res.status(403).send('This is not your task!');
+            }
+            else {
+                req.flash('info', 'Task deleted');
+                res.redirect('/')
+            }
+        })
+        .catch( (err) => {
+            next(err);
+        });
+
 });
 /*marks all the tasks as done*/
 router.post('/alldone', function(req, res, next){
-    var newDate = new Date();
-  Task.updateMany({completed: false}, {completed: true}, {dateCompleted: newDate})
-      .then(()=> {
-      req.flash('info', 'All tasks are done!');
-      res.redirect('/');
-      })
-      .catch((err) => {
-      next(err);
-      })
+
+    Task.update( {_creator: req.user, completed: false}, {completed: true}, {multi: true})
+        .then( (result) => {
+            req.flash('info', 'All tasks are done!');
+            res.redirect('/')
+        })
+        .catch( (err) => {
+            next(err);
+        });
 });
 /*getting a specific page per task*/
 router.get('/task' + '/:_id', function(req, res, next){
-  Task.findById(req.params._id)
-      .then((doc) => {
-    if(doc){
-          res.render('task', {task: doc});
-      }
-      else{
-        next();
-    }
-      })
-    .catch((err) => {
-        var error = new Error('Task not found');
-        error.status = 404;
+    Task.findById(req.params.id).then( (task) => {
+
+        if (!task) {
+            res.status(404).send('Task not found.');
+        }
+
+        // Verify that this task was created by the currently logged in user
+        else if (!task._creator.equals(req.user._id)) {
+            res.status(403).send('This is not your task!');  // 403 Unauthorized
+        }
+
+        else {
+            res.render('task_detail', {task:task})
+        }
+
+    }).catch( (err) => {
         next(err);
     });
 });
 /*delete all the completed task in the completed page*/
 router.post('/deleteDone', function(req, res, next){
-    Task.deleteMany({completed:true})
+    Task.deleteMany({_creator: req.user, completed:true})
         .then(()=>{
-            res.redirect('/');
             req.flash('info', 'All completed tasks deleted');
+            res.redirect('/');
         })
         .catch((err) => {
             next(err);
         })
 });
+
 module.exports = router;
